Move FeedProfile validator into test/support
The validator only runs in CI; promoting it out of app/ avoids shipping
unused code in the runtime image. Test class flattens to FeedProfileValidatorTest.
