--- app/models/feed_profile/validator.rb
@@ -1,86 +1,0 @@
-class FeedProfile
-  # Validates the shape of every entry in FeedProfile::PROFILES.
-  # Invoked at boot from config/initializers/feed_profile_validator.rb so
-  # registry typos surface immediately instead of at runtime.
-  #
-  # Schema mirrors specs/001-smart-feed-creation/contracts/profile_registry.md.
-  module Validator
-    Error = Class.new(StandardError)
-
-    ENTRY_SCHEMA = {
-      "type" => "object",
-      "additionalProperties" => false,
-      "required" => %w[
-        display_name
-        description
-        input_shape
-        depends_on_ai
-        matcher
-        parameter_schema
-        loader
-        processor
-        normalizer
-      ],
-      "properties" => {
-        "display_name" => { "type" => "string", "minLength" => 1, "maxLength" => 80 },
-        "description" => { "type" => "string", "minLength" => 1, "maxLength" => 200 },
-        "input_shape" => { "type" => "string", "enum" => %w[url handle query any] },
-        "depends_on_ai" => { "type" => "boolean" },
-        "matcher" => { "type" => "string", "minLength" => 1 },
-        "parameter_schema" => { "type" => "object" },
-        "loader" => { "$ref" => "#/$defs/stage_entry" },
-        "processor" => { "$ref" => "#/$defs/stage_entry" },
-        "normalizer" => { "$ref" => "#/$defs/stage_entry" },
-        "title_extractor" => { "type" => %w[string null] },
-        "output_schema" => { "type" => %w[object null] }
-      },
-      "$defs" => {
-        "stage_entry" => {
-          "type" => "object",
-          "required" => %w[class config],
-          "properties" => {
-            "class" => { "type" => "string", "minLength" => 1 },
-            "config" => { "type" => "object" }
-          },
-          "additionalProperties" => false
-        }
-      }
-    }.freeze
-
-    def self.validate!(profiles = FeedProfile::PROFILES)
-      schemer = JSONSchemer.schema(ENTRY_SCHEMA)
-      failures = []
-
-      profiles.each do |key, entry|
-        stringified = deep_stringify_symbols(entry)
-        schemer.validate(stringified).each do |err|
-          pointer = err["data_pointer"].to_s
-          failures << "FeedProfile #{key.inspect}#{pointer}: #{err['error']}"
-        end
-
-        if entry[:depends_on_ai] && !entry[:output_schema].is_a?(Hash)
-          failures << "FeedProfile #{key.inspect}: output_schema is required when depends_on_ai is true"
-        end
-      end
-
-      return if failures.empty?
-
-      raise Error, "Invalid FeedProfile registry:\n  - " + failures.join("\n  - ")
-    end
-
-    def self.deep_stringify_symbols(value)
-      case value
-      when Hash
-        value.each_with_object({}) do |(k, v), acc|
-          acc[k.is_a?(Symbol) ? k.to_s : k] = deep_stringify_symbols(v)
-        end
-      when Array
-        value.map { |v| deep_stringify_symbols(v) }
-      when Symbol
-        value.to_s
-      else
-        value
-      end
-    end
-  end
-end
